moving cwl files creation to a new class
